refactor(#5061): Move transpilation logic to a new Transpile class
