+++ eo-maven-plugin/src/main/java/org/eolang/maven/Transpile.java
@@ -1,0 +1,398 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
+ * SPDX-License-Identifier: MIT
+ */
+package org.eolang.maven;
+
+import com.github.lombrozo.xnav.Filter;
+import com.github.lombrozo.xnav.Xnav;
+import com.jcabi.log.Logger;
+import com.jcabi.xml.XML;
+import com.jcabi.xml.XMLDocument;
+import com.yegor256.xsline.Shift;
+import com.yegor256.xsline.StClasspath;
+import com.yegor256.xsline.TrClasspath;
+import com.yegor256.xsline.TrDefault;
+import com.yegor256.xsline.TrJoined;
+import com.yegor256.xsline.Train;
+import com.yegor256.xsline.Xsline;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.Collection;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Function;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+import org.apache.maven.project.MavenProject;
+import org.cactoos.func.StickyFunc;
+import org.eolang.parser.OnDefault;
+import org.eolang.parser.OnDetailed;
+import org.eolang.parser.TrFull;
+
+/**
+ * Transpile.
+ * <p>
+ *     This class transpiles XMIR files into Java source files.
+ *     The goal of this class is to automate the process of transforming EO
+ *     intermediate representations (XMIRs) into Java code, optimize the output,
+ *     and integrate the results into the Maven build lifecycle for further
+ *     processing.
+ * </p>
+ * <p>
+ *     This is one of the final goals that transpiles XMIRs to Java files generated by
+ *     {@link MjAssemble}, {@link MjParse} or {@link MjPull} goals.
+ *     The resulting Java files are stored in the {@link #DIR} directory.
+ *     The intermediate optimized XMIRs are stored in the {@link #PRE} directory.
+ * </p>
+ * @since 0.1
+ * @checkstyle ClassFanOutComplexityCheck (500 lines)
+ */
+final class Transpile {
+
+    /**
+     * The directory where to transpile to.
+     */
+    static final String DIR = "5-transpile";
+
+    /**
+     * The directory where to put pre-transpile files.
+     */
+    static final String PRE = "5-pre-transpile";
+
+    /**
+     * Cache directory for transpiled sources.
+     */
+    static final String CACHE = "transpiled";
+
+    /**
+     * Java extension.
+     */
+    private static final String JAVA = "java";
+
+    /**
+     * Parsing train with XSLs.
+     */
+    static final Train<Shift> TRAIN = new TrFull(
+        new TrJoined<>(
+            new TrClasspath<>(
+                "/org/eolang/maven/transpile/set-locators.xsl",
+                "/org/eolang/maven/transpile/set-original-names.xsl",
+                "/org/eolang/maven/transpile/classes.xsl",
+                "/org/eolang/maven/transpile/tests.xsl",
+                "/org/eolang/maven/transpile/anonymous-to-nested.xsl",
+                "/org/eolang/maven/transpile/package.xsl",
+                "/org/eolang/maven/transpile/attrs.xsl",
+                "/org/eolang/maven/transpile/data.xsl"
+            ).back(),
+            new TrDefault<>(
+                new StClasspath(
+                    "/org/eolang/maven/transpile/to-java.xsl",
+                    String.format("disclaimer %s", new Disclaimer())
+                )
+            )
+        )
+    );
+
+    /**
+     * XMIR sources to transpile.
+     */
+    private final Collection<TjForeign> sources;
+
+    /**
+     * Maven project (used to register generated source roots).
+     */
+    private final MavenProject project;
+
+    /**
+     * Whether to add the generated directory to Maven's compile source roots.
+     */
+    private final boolean addSourcesRoot;
+
+    /**
+     * Target directory.
+     */
+    private final Path targetDir;
+
+    /**
+     * Generated sources directory.
+     */
+    private final Path generatedDir;
+
+    /**
+     * Base cache directory.
+     */
+    private final Path cacheDir;
+
+    /**
+     * Whether caching is enabled.
+     */
+    private final boolean cacheEnabled;
+
+    /**
+     * Plugin version.
+     */
+    private final String version;
+
+    /**
+     * Whether to track transformation steps into intermediate XMIR files.
+     */
+    private final boolean trackTransformationSteps;
+
+    /**
+     * Whether to transpile tests.
+     */
+    private final boolean transpileTests;
+
+    /**
+     * Measured XSL transformation train.
+     */
+    private final Train<Shift> train;
+
+    /**
+     * Constructor.
+     * @param srcs XMIR sources to transpile
+     * @param prj Maven project
+     * @param roots Whether to add generated dir to compile source roots
+     * @param target Target directory
+     * @param generated Generated sources directory
+     * @param cache Base cache directory
+     * @param enabled Whether caching is enabled
+     * @param ver Plugin version string
+     * @param tracking Whether to track transformation steps
+     * @param tests Whether to transpile tests
+     * @param measured Measured XSL transformation train
+     * @checkstyle ParameterNumberCheck (15 lines)
+     */
+    Transpile(
+        final Collection<TjForeign> srcs,
+        final MavenProject prj,
+        final boolean roots,
+        final Path target,
+        final Path generated,
+        final Path cache,
+        final boolean enabled,
+        final String ver,
+        final boolean tracking,
+        final boolean tests,
+        final Train<Shift> measured
+    ) {
+        this.sources = srcs;
+        this.project = prj;
+        this.addSourcesRoot = roots;
+        this.targetDir = target;
+        this.generatedDir = generated;
+        this.cacheDir = cache;
+        this.cacheEnabled = enabled;
+        this.version = ver;
+        this.trackTransformationSteps = tracking;
+        this.transpileTests = tests;
+        this.train = measured;
+    }
+
+    /**
+     * Run transpilation of all sources.
+     * @throws IOException If any issues with I/O
+     */
+    @SuppressWarnings("PMD.UnnecessaryLocalRule")
+    void exec() throws IOException {
+        final long begin = System.currentTimeMillis();
+        final int saved = new Threaded<>(
+            this.sources,
+            this::transpiled
+        ).total() + new PackageInfos(this.generatedDir).create();
+        Logger.info(
+            this, "Transpiled %d XMIRs, created %d Java files in %[file]s",
+            this.sources.size(), saved, this.generatedDir
+        );
+        if (this.addSourcesRoot) {
+            this.project.addCompileSourceRoot(this.generatedDir.toAbsolutePath().toString());
+            Logger.info(
+                this, "The directory added to Maven 'compile-source-root': %[file]s",
+                this.generatedDir
+            );
+            final String gtests = this.generatedDir.getParent().resolve(
+                "generated-test-sources"
+            ).toAbsolutePath().toString();
+            this.project.addTestCompileSourceRoot(gtests);
+            Logger.info(
+                this, "The directory added to Maven 'test-compile-source-root': %[file]s",
+                gtests
+            );
+        }
+        Logger.info(
+            this,
+            "Transpilation took %[ms]s in total",
+            System.currentTimeMillis() - begin
+        );
+    }
+
+    /**
+     * Transpile a single tojo.
+     * @param tojo Tojo that should be transpiled
+     * @return Number of generated Java files
+     * @throws IOException If any issues with I/O
+     */
+    @SuppressWarnings("PMD.UnnecessaryLocalRule")
+    private int transpiled(final TjForeign tojo) throws IOException {
+        final Path source = tojo.xmir();
+        final XML xmir = new XMLDocument(source);
+        final Path base = this.targetDir.resolve(Transpile.DIR);
+        final Path target = new Place(
+            new OnDetailed(new OnDefault(xmir), source).get()
+        ).make(base, MjAssemble.XMIR);
+        final Supplier<String> hsh = new TojoHash(tojo);
+        final AtomicBoolean rewrite = new AtomicBoolean(false);
+        final Function<XML, XML> transform = this.transpilation(source);
+        final Path cdir = this.cacheDir.resolve(Transpile.CACHE);
+        final Path tail = base.relativize(target);
+        if (this.cacheEnabled) {
+            new ConcurrentCache(
+                new Cache(
+                    new CachePath(cdir, this.version, hsh.get()),
+                    src -> {
+                        rewrite.compareAndSet(false, true);
+                        final long start = System.currentTimeMillis();
+                        final String res = transform.apply(xmir).toString();
+                        Logger.debug(
+                            this,
+                            "Transpiled %[file]s (%s) to %[file]s (%s) in %[ms]s (cache miss), version: %s, hash: %s, tail: %s, cache enabled: %b, cache dir: %[file]s",
+                            source,
+                            Transpile.info(source),
+                            target,
+                            Transpile.info(target),
+                            System.currentTimeMillis() - start,
+                            this.version,
+                            hsh.get(),
+                            tail,
+                            this.cacheEnabled,
+                            cdir
+                        );
+                        return res;
+                    }
+                )
+            ).apply(source, target, tail);
+        } else {
+            rewrite.compareAndSet(false, true);
+            new Saved(transform.apply(xmir).toString(), target).value();
+        }
+        return this.javaGenerated(rewrite.get(), target, hsh.get());
+    }
+
+    /**
+     * File info for logging.
+     * @param info Path to file
+     * @return Info string
+     * @throws IOException If fails
+     */
+    private static String info(final Path info) throws IOException {
+        final String res;
+        if (Files.exists(info)) {
+            res = Files.getLastModifiedTime(info).toString();
+        } else {
+            res = "Not exists yet";
+        }
+        return res;
+    }
+
+    /**
+     * Build XSL transformation function for a source file.
+     * If {@code trackTransformationSteps} is {@code true} - creates a new {@link Xsline}
+     * for every XMIR in purpose of thread safety.
+     * @param source Path to source XMIR
+     * @return XSL transformation function
+     */
+    private Function<XML, XML> transpilation(final Path source) {
+        final Function<XML, XML> func;
+        if (this.trackTransformationSteps) {
+            func = xml -> new Xsline(
+                new TrSpy(
+                    this.train,
+                    new StickyFunc<>(
+                        doc -> new Place(
+                            new OnDetailed(new OnDefault(doc), source).get()
+                        ).make(this.targetDir.resolve(Transpile.PRE), "")
+                    )
+                )
+            ).pass(xml);
+        } else {
+            func = new Xsline(this.train)::pass;
+        }
+        return func;
+    }
+
+    /**
+     * Generate java files and count them.
+     * @param rewrite Rewrite .java files even if they exist
+     * @param target Full target path to XMIR after transpilation optimizations
+     * @param hsh Tojo hash
+     * @return Amount of generated .java files
+     * @throws IOException If fails to save files
+     */
+    @SuppressWarnings("PMD.UnnecessaryLocalRule")
+    private int javaGenerated(
+        final boolean rewrite,
+        final Path target,
+        final String hsh
+    ) throws IOException {
+        final long begin = System.currentTimeMillis();
+        final AtomicInteger saved = new AtomicInteger(0);
+        if (Files.exists(target)) {
+            final Xnav object = new Xnav(target).element("object");
+            final Collection<Xnav> classes = object.elements(Filter.withName("class"))
+                .collect(Collectors.toList());
+            final boolean atom = object.path("/object/o/o[@name='λ']").findAny().isPresent();
+            for (final Xnav clazz : classes) {
+                final String jname = clazz.attribute("java-name").text().get();
+                if (!atom || jname.endsWith("Test")) {
+                    final Path tgt = new Place(jname).make(
+                        this.generatedDir, Transpile.JAVA
+                    );
+                    final Footprint java = new FpJavaGenerated(
+                        clazz, new FileGenerationReport(saved, tgt, target)
+                    );
+                    new JavaPlaced(
+                        new FpIfReleased(
+                            hsh,
+                            new FpAppliedWithCache(
+                                java,
+                                this.cached(hsh, jname),
+                                new RewritePolicy(rewrite, tgt),
+                                this.cacheEnabled
+                            ),
+                            java
+                        ),
+                        tgt,
+                        this.generatedDir
+                    ).exec(clazz, this.transpileTests);
+                }
+            }
+            Logger.debug(
+                this,
+                "Generated %d Java files from %[file]s in %[ms]s",
+                saved.get(), target, System.currentTimeMillis() - begin
+            );
+        }
+        return saved.get();
+    }
+
+    /**
+     * Cached path supplier for a generated Java file.
+     * @param hsh Hash
+     * @param jname Java class name
+     * @return Supplier of cached path
+     */
+    private Supplier<Path> cached(final String hsh, final String jname) {
+        return new CachePath(
+            this.cacheDir.resolve(Transpile.CACHE),
+            this.version,
+            hsh,
+            this.generatedDir.relativize(
+                new Place(jname).make(
+                    this.generatedDir, Transpile.JAVA
+                )
+            )
+        );
+    }
+}
