+++ eo-maven-plugin/src/test/java/org/eolang/maven/MjTranspileTest.java
@@ -139,7 +139,7 @@ final class MjTranspileTest {
                 .result(),
             Matchers.not(
                 Matchers.allOf(
-                    Matchers.hasKey(String.format("target/%s/foo/x/main.xmir", MjTranspile.DIR)),
+                    Matchers.hasKey(String.format("target/%s/foo/x/main.xmir", Transpile.DIR)),
                     Matchers.hasKey("target/generated/EOcom/EOexample/EOfoo.java")
                 )
             )
@@ -224,7 +224,7 @@ final class MjTranspileTest {
             .result();
         final Path java = res.get(this.compiled);
         final Path xmir = maven.targetPath().resolve(
-            String.format("%s/foo/x/main.xmir", MjTranspile.DIR)
+            String.format("%s/foo/x/main.xmir", Transpile.DIR)
         );
         MatcherAssert.assertThat(
             "The Java file should exist after transpile",
